Update for complete python2.7 compatibility
Had issues with the mixture of some Python2.7 and Python3 packages,
so I made the project completely depdendent only on Python2.7 with
the following changes:

I chose to go with the Python2.7 requirement over Python3 because
the `mechanize` library has no Python3 compatible version.

- Use Python2.7 urllib and urlparse libraries for compatible
functionality of the python3 urllib merged library
- Updated all `input` methods to `raw_input` to fix type errors
for user input
- Change test from `builtins.zip` to `future_builtins.zip` which
was added to Python2.7 since the library was coming to Python3
- Correct misleading error message about version of python noting
Python2.7 requirement

I have tested all changes in a clean Python2.7 virtualenv and
to my knowledge all functionality is intact with the changes.