Improve and document dict replacement function.
The function replace_value has a third strategy defining value
that is per default causing inplace changes, but can be used
via copy.copy or copy.deepcopy to work on shallow or deep clones
of the parameter dict where the replacement should take place.

As small optimzation for small and large dicts the value dictview
object is used as a filter before iterating ofer the keys.